Move ObjectParser and ObjectSerializer to dedicated package
