+++ bundles/org.openhab.core/src/main/java/org/openhab/core/converter/ObjectSerializer.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.core.common;
+package org.openhab.core.converter;
 
 import java.io.OutputStream;
